Use sync strategy to determine blaze info binary path
Adds utility methods to BuildSystem for returning a default BuildInvoker ands binary path for syncs. If the sync strategy is PARALLEL, the methods returns parallel invoker/path, if it exists. In all other cases, the default is returned.

This ensures that initial invocations of `info` and `query` use the parallel invoker binary when applicable. When sync strategy is DECIDE_AUTOMATICALLY, the local invoker is used through the query invocation.

PiperOrigin-RevId: 451024320